♻️ 리팩터: TransportActivityService CO₂ 절감량 계산 로직 개선

+++ src/main/java/com/penglobe/server/dto/RegionDTO.java
@@ -2,14 +2,14 @@ package com.penglobe.server.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter @Setter
 @NoArgsConstructor @AllArgsConstructor @Builder
 public class RegionDTO {
 
     private Integer regionId; // 지역 아이디
     private String regionName; // 지역 이름
-    private Long totalSaving; // 총 절감량
-
-
+    private BigDecimal totalSaving; // 총 절감량
 
 }
